Implement a naive version of Kociemba's algorithm
Implement a very basic Rubik's cube solver that takes as input a scrambe
string and returns a solution string in face turn notation.

This implementation impelents the core components of Kociemba's algorithm,
but leaves out some of the more complex optimizations that the original
has. We do not compute a perfect heuristic for Phase1 and we do not deal
with cube symmetries at all. It still manages decent performance without
them, finding solutions in a few seconds most of the time.